+++ packages/rangeslider/CHANGELOG.md
@@ -1,5 +1,14 @@
 # `rangeslider` Changelog
 
+## [v1.1.0] 15th November 2021
+
+- Added new `fs-rangefilter-formatdisplay="true"` attribute.
+  Now, by default all `fs-rangefilter-element="display-value"` elements will display the numeric value without formatting.
+  If the `formatdisplay` attribute is added to the `wrapper` element, the `display-value` elements will display a formatted value (formatting defaults to the user's country conventions, some countries use commas and other countries use dots).
+- Added an extra check to make sure the Handles' start values are adjusted to the `step` increment coefficient.
+- Added an extra rule to set the Handles' start value to the `minRange` or `maxRange` if the provided value doesn't match the range. When this happens, a custom `Debug` message will be displayed to warn the user when using `debugMode`.
+- Added an extra `Debug` alert that warns the user when the provided `step` value doesn't fit the provided `[min,max]` range.
+
 ## [v1.0.1] 15th November 2021
 
 - Made sure the track is set to `position: relative` and the Handles & Fill have the correct `position: absolute` values.